Travis YML file updated into the repo

+++ rmdock.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
 for FILE in `docker images | grep "none"`;do
-`docker rmi $FILE`;
+  if [ $FILE != "juniornsa/todo-*" ]; then
+	  echo "file doesnt have the name on it , must be deleted"
+	  `docker rmi $FILE`
+  fi
 done
